+++ examples/mouse_picking.rs
@@ -1,20 +1,13 @@
-use bevy::{prelude::*, window::PresentMode};
+//! This example will show you how to use your mouse cursor as a ray casting source, cast into the
+//! scene, intersect a mesh, and mark the intersection with the built in debug cursor. If you are
+//! looking for a more fully-featured mouse picking plugin, try out bevy_mod_picking.
 
+use bevy::prelude::*;
 use bevy_mod_raycast::{prelude::*, print_intersections};
-
-// This example will show you how to use your mouse cursor as a ray casting source, cast into the
-// scene, intersect a mesh, and mark the intersection with the built in debug cursor. If you are
-// looking for a more fully-featured mouse picking plugin, try out bevy_mod_picking.
 
 fn main() {
     App::new()
-        .add_plugins(DefaultPlugins.set(WindowPlugin {
-            primary_window: Some(Window {
-                present_mode: PresentMode::AutoNoVsync, // Reduces input lag.
-                ..default()
-            }),
-            ..default()
-        }))
+        .add_plugins(DefaultPlugins)
         // The DefaultRaycastingPlugin bundles all the functionality you might need into a single
         // plugin. This includes building rays, casting them, and placing a debug cursor at the
         // intersection. For more advanced uses, you can compose the systems in this plugin however
@@ -36,7 +29,7 @@ fn main() {
 /// This is a unit struct we will use to mark our generic `RaycastMesh`s and `RaycastSource` as part
 /// of the same group, or "RaycastSet". For more complex use cases, you might use this to associate
 /// some meshes with one ray casting source, and other meshes with a different ray casting source."
-#[derive(Clone, Reflect)]
+#[derive(Reflect)]
 struct MyRaycastSet;
 
 // Update our `RaycastSource` with the current cursor position every frame.