+++ examples/reflecting_laser.rs
@@ -1,27 +1,24 @@
-//! This example demonstrates how to use the [`Raycast`] system param in queries to run raycasts
-//! on-demand, in an immediate mode style. Unlike using a [`RaycastSource`]
+//! This example demonstrates how to use the [`Raycast`] system param to chain multiple raycasts and
+//! bounce off of surfaces.
 
 use bevy::prelude::*;
 use bevy_mod_raycast::prelude::*;
 
 fn main() {
     App::new()
-        .add_plugins((
-            DefaultPlugins,
-            DefaultRaycastingPlugin::<MyRaycastSet>::default(),
-        ))
+        .add_plugins((DefaultPlugins, DefaultRaycastingPlugin::<Laser>::default()))
         .add_systems(Startup, setup_scene)
-        .add_systems(Update, immediate_mode_raycast)
+        .add_systems(Update, bouncing_raycast)
         .run();
 }
 
 const MAX_BOUNCES: usize = 128;
-const LASER_MOVE_SPEED: f32 = 0.05;
+const LASER_MOVE_SPEED: f32 = 0.005;
 
-#[derive(Reflect, Clone)]
-struct MyRaycastSet;
+#[derive(Reflect)]
+struct Laser;
 
-fn immediate_mode_raycast(raycast: Raycast<MyRaycastSet>, mut gizmos: Gizmos, time: Res<Time>) {
+fn bouncing_raycast(raycast: Raycast<Laser>, mut gizmos: Gizmos, time: Res<Time>) {
     let t = (time.elapsed_seconds() * LASER_MOVE_SPEED).sin() * std::f32::consts::PI;
     let mut ray_pos = Vec3::new(t.sin(), (3.0 * t).cos() * 0.5, t.cos()) * 3.0;
     let mut ray_dir = (-ray_pos).normalize();
@@ -32,11 +29,11 @@ fn immediate_mode_raycast(raycast: Raycast<MyRaycastSet>, mut gizmos: Gizmos, ti
 
     for i in 0..MAX_BOUNCES {
         let ray = Ray3d::new(ray_pos, ray_dir);
-        if let Some((_, hit)) = raycast.cast_ray(ray, false).values().next() {
-            intersections.push((
-                hit.position(),
-                Color::rgba(1.0, 0.0, 0.0, 1.0 - i as f32 / MAX_BOUNCES as f32),
-            ));
+        if let Some((_, hit)) = raycast.cast_ray(ray, false).first() {
+            let a = 0.2 + 0.8 * (1.0 - i as f32 / MAX_BOUNCES as f32);
+            let color = Color::rgba(1.0, 0.0, 0.0, a);
+            intersections.push((hit.position(), color));
+            gizmos.sphere(hit.position(), Quat::IDENTITY, 0.02, color);
             // reflect the ray
             let proj = (ray_dir.dot(hit.normal()) / hit.normal().dot(hit.normal())) * hit.normal();
             ray_dir = (ray_dir - 2.0 * proj).normalize();
@@ -54,7 +51,7 @@ fn setup_scene(
     mut meshes: ResMut<Assets<Mesh>>,
     mut materials: ResMut<Assets<StandardMaterial>>,
 ) {
-    commands.insert_resource(RaycastPluginState::<MyRaycastSet>::default().with_debug_cursor());
+    commands.insert_resource(RaycastPluginState::<Laser>::default().with_debug_cursor());
     commands.spawn(DirectionalLightBundle {
         transform: Transform::from_rotation(Quat::from_euler(EulerRot::XYZ, 1.0, 1.0, 1.0)),
         ..Default::default()
@@ -64,7 +61,7 @@ fn setup_scene(
             transform: Transform::from_xyz(7.0, 5.0, 9.0).looking_at(Vec3::Y * -1.0, Vec3::Y),
             ..default()
         },
-        RaycastSource::<MyRaycastSet>::new(), // Designate the camera as our source
+        RaycastSource::<Laser>::new(), // Designate the camera as our source
     ));
     commands.spawn((
         PbrBundle {
@@ -73,7 +70,8 @@ fn setup_scene(
             transform: Transform::from_scale(Vec3::splat(6.0)),
             ..default()
         },
-        RaycastMesh::<MyRaycastSet>::default(),
+        RaycastMesh::<Laser>::default(),
+        // Without this, raycasts would shoot straight out from the inside of the cube.
         bevy_mod_raycast::NoBackfaceCulling,
     ));
 }